Migrate to serde 1.0
