+++ Cargo.toml
@@ -8,21 +8,17 @@ authors = [ "Kirill Paltsev <kpaltsev@avito.ru>",
 description = "Microservice for image processing"
 license = "MIT"
 
-build = "build.rs"
-
 [dependencies]
-image = "~0.10"
-url="~1.2"
-serde = "~0.8"
-serde_json = "~0.8"
+image = "^0.10"
+url="^1.2"
+serde = "^1.0"
+serde_derive = "^1.0"
+serde_json = "^1.0"
 "quick-error"="~1"
 "scoped-pool" = "~1"
-clap = "~2.16"
-hyper="~0.9"
-regex="~0.1"
-liquid="~0.8"
-multipart="~0.8"
-
-[build-dependencies]
-serde_codegen = "0.8"
+clap = "^2"
+hyper="^0.9"
+regex="^0.1"
+liquid="^0.8"
+multipart="^0.8"
