+++ src/config.rs
@@ -3,7 +3,31 @@ use std::collections::HashMap;
 
 use actions::*;
 
-include!(concat!(env!("OUT_DIR"), "/config_types.rs"));
+//#include!(concat!(env!("OUT_DIR"), "/config_types.rs"));
+#[derive(Serialize, Deserialize, Clone)]
+pub struct Config {
+    pub presets: HashMap<String, Preset>,
+}
+
+#[derive(Serialize, Deserialize, Clone)]
+pub struct Preset {
+    pub name: String,
+    pub tasks: Vec<Task>,
+}
+
+#[derive(Serialize, Deserialize, Clone)]
+pub struct Task {
+    pub name: String,
+
+    #[serde(default = "Vec::new")]
+    #[serde(skip_deserializing)]
+    #[serde(skip_serializing)]
+    #[serde(rename = "actions_box")]
+    pub actions: Vec<Action>,
+    #[serde(rename = "actions")]
+    pub actions_raw: Vec<Vec<String>>,
+    pub url_template: String,
+}
 
 impl Task {
     fn init(&mut self) -> Result<(), ActionError> {
