+++ src/main.rs
@@ -1,8 +1,11 @@
 extern crate image;
 extern crate url;
-#[macro_use] extern crate quick_error;
+#[macro_use]
+extern crate quick_error;
 extern crate serde;
 extern crate serde_json;
+#[macro_use]
+extern crate serde_derive;
 extern crate scoped_pool;
 extern crate clap;
 extern crate hyper;
@@ -33,26 +36,23 @@ fn main() {
 
     let matches = App::new("Avito gravure - The best image service ever")
         .arg(Arg::with_name("config")
-             .short("c")
-             .long("config")
-             .value_name("FILE")
-             .help("Sets a custom config file")
-             .takes_value(true)
-            )
+                 .short("c")
+                 .long("config")
+                 .value_name("FILE")
+                 .help("Sets a custom config file")
+                 .takes_value(true))
         .arg(Arg::with_name("listen")
-             .short("l")
-             .long("listen")
-             .value_name("HOST:[PORT]")
-             .help("Listening parameters")
-             .takes_value(true)
-            )
+                 .short("l")
+                 .long("listen")
+                 .value_name("HOST:[PORT]")
+                 .help("Listening parameters")
+                 .takes_value(true))
         .arg(Arg::with_name("threads")
-             .short("n")
-             .long("threads")
-             .value_name("NUM")
-             .help("number of thread")
-             .takes_value(true)
-            )
+                 .short("n")
+                 .long("threads")
+                 .value_name("NUM")
+                 .help("number of thread")
+                 .takes_value(true))
         .get_matches();
 
     let config = matches.value_of("config").unwrap_or("config_test.json");
@@ -61,13 +61,17 @@ fn main() {
     config.init().unwrap();
 
     let listen = matches.value_of("listen").unwrap_or("0.0.0.0:4444");
-    let threads = matches.value_of("threads").unwrap_or("8").parse().unwrap();
+    let threads = matches
+        .value_of("threads")
+        .unwrap_or("8")
+        .parse()
+        .unwrap();
 
     let (job_s, job_r) = mpsc::channel();
     let server = rest::GravureServer::new(config, "upload".to_owned(), job_s);
 
     let queue = Queue::new(threads, job_r);
-    let handle = thread::spawn(move || {queue.run();});
+    let handle = thread::spawn(move || { queue.run(); });
     Server::http(listen).unwrap().handle(server).unwrap();
     handle.join().unwrap();
 }
